refactor: 일부 fsd 아키텍처 적용
