+++ .github/CONTRIBUTING.md
@@ -46,8 +46,6 @@ _Pull requests, bug reports, and all other forms of contribution are welcomed an
 - `src` - main source.
   - `contentScript` - scripts and components to be injected as `content_script`
   - `background` - scripts for background.
-  - `components` - auto-imported Vue components that are shared in popup and options page.
-  - `styles` - styles shared in popup and options page
   - `assets` - assets used in Vue components
   - `manifest.ts` - manifest for the extension.
 - `extension` - extension package root.
