+++ .gitignore
@@ -12,6 +12,5 @@ dist
 dist-ssr
 extension/manifest.json
 node_modules
-src/auto-imports.d.ts
-src/components.d.ts
+src/types
 .eslintcache
