+++ src/entities/storage/index.ts
@@ -1,0 +1,1 @@
+export { useStorageDemo, useStorageTheme } from './model/storage';
