+++ src/entities/storage/model/storage.ts
@@ -1,0 +1,9 @@
+import { useWebExtensionStorageLocal } from './useWebExtensionStorage';
+
+export const useStorageDemo = () => {
+  return useWebExtensionStorageLocal('webext-demo', 'Storage Demo');
+};
+
+export const useStorageTheme = () => {
+  return useWebExtensionStorageLocal('theme', 'dark');
+};
