--- src/global.d.ts
@@ -1,3 +1,0 @@
-declare const __DEV__: boolean;
-/** Extension name, defined in packageJson.name */
-declare const __NAME__: string;
