--- src/logic/storage.ts
@@ -1,5 +1,0 @@
-import { useWebExtensionStorageLocal } from '~/shared/model/useWebExtensionStorage';
-
-export function useStorageDemo() {
-  return useWebExtensionStorageLocal('webext-demo', 'Storage Demo');
-}
