+++ src/options/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import '../styles';
+import '../shared/styles';
 import App from './Options';
 
 const container = document.getElementById('app');
