+++ src/shared/lib/env.ts
