+++ src/sidebar/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 
-import '../styles';
+import '../shared/styles';
 import App from './Sidebar';
 
 const container = document.getElementById('app');
